+++ lib/helpers/snapshot.sh
@@ -30,6 +30,8 @@ targets.skills.dest
 targets.skills.inline_into_agents
 targets.commands.dest
 targets.commands.format
+targets.commands.as_skills
+targets.commands.inline_into_agents
 targets.subagents.dest
 targets.subagents.format
 targets.settings.source